rename instrs.rs -> executor.rs
